Video progress bar doesn't stop when a video is paused in the Help Section https://jira.ccpgames.com/browse/EO-15314
Video controller gets all confused when it's started with an audio sink for a video without an audio: in that case audio sink is reset after video's metadata gets parsed, and the simple timer is supposed to take over video timing from an audio sink. But the timer is already in a wrong state. Fix: inform the simple timer of all pause/resume, etc. changes regardless if there is an audio sink.

[git-p4: depot-paths = "//eve/branches/development/MAINLINE/carbon/src/VideoPlayer/": change = 1897088]